backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
langchain==0.1.16
langchain-community==0.2.16
langchain-core==0.2.38
langsmith==0.1.98
chromadb==0.4.24
python-dotenv==1.0.1